requirements.txt
Flask
gunicorn
praat-parselmouth
numpy
matplotlib
seaborn